# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: API для управления банковскими картами без аутентификации
  version: 1.0.0
  contact:
    name: Bank Development Team
    email: dev@bank.com

servers:
  - url: http://localhost:8080
    description: Development server

components:
  schemas:
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - email
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]
    
    UpdateUserRequest:
      type: object
      required:
        - username
        - email
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, ADMIN]
    
    BankCardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        cardholderName:
          type: string
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
        balance:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ownerId:
          type: integer
          format: int64
        ownerUsername:
          type: string
    
    CreateCardRequest:
      type: object
      required:
        - cardholderName
        - ownerId
        - initialBalance
      properties:
        cardholderName:
          type: string
          maxLength: 100
        ownerId:
          type: integer
          format: int64
        initialBalance:
          type: number
          format: decimal
          minimum: 0
    
    UpdateCardStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
    
    TransactionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCardMasked:
          type: string
        toCardMasked:
          type: string
        amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, CANCELLED]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
    
    CreateTransactionRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: decimal
          minimum: 0.01
        description:
          type: string

paths:
  /api/users:
    get:
      summary: Получение списка всех пользователей
      tags: [Users]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            default: id
        - name: sortDir
          in: query
          schema:
            type: string
            default: asc
            enum: [asc, desc]
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
    
    post:
      summary: Создание нового пользователя
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Неверные данные запроса
  
  /api/users/{id}:
    get:
      summary: Получение пользователя по ID
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Пользователь не найден
    
    put:
      summary: Обновление пользователя
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Пользователь успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Неверные данные запроса
        '404':
          description: Пользователь не найден
    
    delete:
      summary: Удаление пользователя
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
  
  /api/users/role/{role}:
    get:
      summary: Получение пользователей по роли
      tags: [Users]
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
            enum: [USER, ADMIN]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список пользователей по роли
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
  
  /api/users/search:
    get:
      summary: Поиск пользователей
      tags: [Users]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
  
  /api/cards:
    get:
      summary: Получение списка всех карт
      tags: [Bank Cards]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            default: id
        - name: sortDir
          in: query
          schema:
            type: string
            default: desc
            enum: [asc, desc]
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BankCardDto'
    
    post:
      summary: Создание новой карты
      tags: [Bank Cards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardDto'
        '400':
          description: Неверные данные запроса
  
  /api/cards/{id}:
    get:
      summary: Получение карты по ID
      tags: [Bank Cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardDto'
        '404':
          description: Карта не найдена
    
    put:
      summary: Обновление статуса карты
      tags: [Bank Cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardStatusRequest'
      responses:
        '200':
          description: Статус карты обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankCardDto'
        '404':
          description: Карта не найдена
    
    delete:
      summary: Удаление карты
      tags: [Bank Cards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Карта успешно удалена
        '404':
          description: Карта не найдена
  
  /api/cards/user/{userId}:
    get:
      summary: Получение карт пользователя
      tags: [Bank Cards]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED]
        - name: searchTerm
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BankCardDto'
  
  /api/cards/status/{status}:
    get:
      summary: Получение карт по статусу
      tags: [Bank Cards]
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список карт по статусу
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BankCardDto'
  
  /api/cards/expired:
    get:
      summary: Получение просроченных карт
      tags: [Bank Cards]
      responses:
        '200':
          description: Список просроченных карт
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankCardDto'
  
  /api/cards/low-balance:
    get:
      summary: Получение карт с низким балансом
      tags: [Bank Cards]
      parameters:
        - name: minBalance
          in: query
          schema:
            type: number
            default: 100
      responses:
        '200':
          description: Список карт с низким балансом
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankCardDto'
  
  /api/transactions:
    get:
      summary: Получение списка всех транзакций
      tags: [Transactions]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sortBy
          in: query
          schema:
            type: string
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            default: desc
            enum: [asc, desc]
      responses:
        '200':
          description: Список транзакций
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionDto'
  
  /api/transactions/transfer:
    post:
      summary: Перевод между картами
      tags: [Transactions]
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '200':
          description: Перевод выполнен успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '400':
          description: Неверные данные запроса
  
  /api/transactions/{id}:
    get:
      summary: Получение транзакции по ID
      tags: [Transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Транзакция найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '404':
          description: Транзакция не найдена
  
  /api/transactions/user/{userId}:
    get:
      summary: История транзакций пользователя
      tags: [Transactions]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: История транзакций
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionDto'
  
  /api/transactions/status/{status}:
    get:
      summary: Получение транзакций по статусу
      tags: [Transactions]
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [PENDING, COMPLETED, FAILED, CANCELLED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список транзакций по статусу
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionDto'
  
  /api/transactions/date-range:
    get:
      summary: Получение транзакций по диапазону дат
      tags: [Transactions]
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список транзакций по диапазону дат
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionDto'
  
  /api/transactions/user/{userId}/all:
    get:
      summary: Получение всех транзакций пользователя
      tags: [Transactions]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Все транзакции пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDto'

tags:
  - name: Users
    description: Управление пользователями
  - name: Bank Cards
    description: Управление банковскими картами
  - name: Transactions
    description: Операции с транзакциями