# src/main/java/com/example/bankcards/dto/CreateTransactionRequest.java
package com.example.bankcards.dto;

import java.math.BigDecimal;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CreateTransactionRequest {
    @NotNull(message = "From card ID is required")
    @Positive(message = "From card ID must be positive")
    private Long fromCardId;
    
    @NotNull(message = "To card ID is required")
    @Positive(message = "To card ID must be positive")
    private Long toCardId;
    
    @NotNull(message = "Amount is required")
    @DecimalMin(value = "0.01", message = "Amount must be greater than 0")
    private BigDecimal amount;
    
    private String description;
}